file_logger middleware, usage example and overal cleanup
